requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asttokens==3.0.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest_asyncio==1.6.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==5.9.0
pure_eval==0.2.3
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==305.1
pyzmq==24.0.1
requests==2.32.3
scikit-learn==1.6.0
scipy==1.14.1
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
stack_data==0.6.3
swarmauri==0.5.2
swarmauri-core==0.5.2
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.44.0
zipp==3.21.0